Init procedure now hooked on getty command (was on /sbin/init)
Fixes #20.

The init procedure of debootstick is now hooked on the getty command
instead of /sbin/init. As a result, this procedure is now called
when the OS is fully booted.
This fixes LVM command failures or long timeouts occuring when the OS
is not booted (no udev daemon, etc).

Moreover, this procedure is now called when user should start to
interact (login) with the system. So it seems a very suitable place
to implement user interaction: ask for root password (cf. option
--config-root-password-first-boot), or allow the user to cancel the
migration process.