installer: workaround for used logical volumes on the target
vgchange fails to deactivate an LV if it finds out that the filesystem
on that LV is in use. Apparently, there is no other way around it then
to remove the signature from the partition and then ask the user to
reboot and try again.

NB: We can also reboot automatically, but normally the user has to pick
again the USB as the boot device. Better then to ask the user to reboot
himself.